# src/nexus/routes/cortex.ts
/**
 * Cortex Feedback Routes
 * API endpoints for cortex feedback, session context, and memory management
 * Exposes services from SessionContextService, ProviderFeedbackEngine, and MemoryAutoFiller
 *
 * @version 2.3.0 - Real metrics and pause/resume implementation
 */

import { Router, Request, Response } from 'express';
import { cortex } from '../../cortex/index.js';
import { precog } from '../../precog/index.js';
import { bus } from '../../core/event-bus.js';

const router = Router();

/**
 * GET /api/v1/cortex/status
 * Get current cortex status for monitoring
 */
router.get('/status', (_req: Request, res: Response) => {
  try {
    const activeProvider = cortex.providerFeedbackEngine.getActiveProvider();
    // Use session context metadata for pause state
    const session = cortex.sessionContextService.getOrCreateSession('default');
    const isPaused = session.metadata['cortex:paused'] === true;

    // Get real metrics from exploration and feedback engines
    const explorationStats = cortex.explorationEngine.getStatistics();
    const feedback = cortex.providerFeedbackEngine.getSerializableFeedback();

    // Calculate real complexity from active experiments and provider load
    const activeExperiments = explorationStats.active || 0;
    const successRate = explorationStats.successRate || 0;
    const avgConfidence = explorationStats.averageConfidence || 0;

    // Complexity: 0-1 based on active work and success rate
    const complexity = Math.min(
      1,
      activeExperiments * 0.2 + (1 - successRate) * 0.3 + (avgConfidence > 0 ? 0.2 : 0)
    );

    // Real bias score from feedback data - use successRate as proxy
    const avgSuccessRate =
      feedback.length > 0
        ? feedback.reduce((acc, p) => acc + (p.successRate || 0), 0) / feedback.length
        : 0.9;
    const biasScore = Math.max(0.01, (1 - avgSuccessRate) * 0.1);

    // Real efficiency from success rate
    const efficiency = successRate > 0 ? successRate : 0.85;

    // Stress based on active experiments and errors
    const stress = Math.min(1, activeExperiments * 0.15 + (explorationStats.rejected || 0) * 0.05);

    const isBusy = !!activeProvider || activeExperiments > 0;

    res.json({
      ok: true,
      ready: !isPaused,
      isPaused,
      isBusy,
      activeProvider: activeProvider ? activeProvider.name : null,
      taskHeadline: isPaused ? 'Paused' : isBusy ? 'Processing request...' : 'Idle',
      complexity: Number(complexity.toFixed(3)),
      metrics: {
        bias: Number(biasScore.toFixed(4)),
        efficiency: Number(efficiency.toFixed(3)),
        stress: Number(stress.toFixed(3)),
        activeExperiments,
        totalExplorations: explorationStats.totalExplorations || 0,
      },
    });
  } catch (e) {
    console.error('[Cortex Routes] Status fetch error:', e);
    res.status(500).json({
      ok: false,
      error: 'Failed to fetch cortex status',
    });
  }
});

/**
 * POST /api/v1/cortex/test-provider
 * Test a specific provider
 */
router.post('/test-provider', async (req: Request, res: Response) => {
  try {
    const { provider } = req.body;
    if (!provider) {
      res.status(400).json({ ok: false, error: 'Provider required' });
      return;
    }

    const startTime = Date.now();
    const result = await precog.providers.generate({
      prompt: "Test connection. Respond with 'OK'.",
      provider,
      taskType: 'test',
    });
    const latency = Date.now() - startTime;

    res.json({
      ok: true,
      result: {
        provider: result.provider,
        content: result.content,
        latency,
      },
    });
  } catch (e) {
    console.error('[Cortex Routes] Provider test error:', e);
    res.status(500).json({
      ok: false,
      error: e instanceof Error ? e.message : 'Test failed',
    });
  }
});

/**
 * POST /api/v1/cortex/pause
 * Pause cortex processing
 */
router.post('/pause', (_req: Request, res: Response) => {
  try {
    // Publish pause event that Executive listens for
    bus.publish('nexus', 'nexus:intervention:pause', {
      requestedAt: new Date().toISOString(),
      reason: 'user_requested',
    });

    // Mark cortex state as paused using session metadata
    const session = cortex.sessionContextService.getOrCreateSession('default');
    session.metadata['cortex:paused'] = true;

    res.json({
      ok: true,
      status: 'paused',
      message: 'Cortex processing paused. Active tasks will complete but no new tasks will start.',
    });
  } catch (e) {
    console.error('[Cortex Routes] Pause error:', e);
    res.status(500).json({
      ok: false,
      error: 'Failed to pause cortex',
    });
  }
});

/**
 * POST /api/v1/cortex/resume
 * Resume cortex processing
 */
router.post('/resume', (_req: Request, res: Response) => {
  try {
    // Publish resume event that Executive listens for
    bus.publish('nexus', 'nexus:intervention:resume', {
      requestedAt: new Date().toISOString(),
    });

    // Mark cortex state as running using session metadata
    const session = cortex.sessionContextService.getOrCreateSession('default');
    session.metadata['cortex:paused'] = false;

    res.json({
      ok: true,
      status: 'running',
      message: 'Cortex processing resumed.',
    });
  } catch (e) {
    console.error('[Cortex Routes] Resume error:', e);
    res.status(500).json({
      ok: false,
      error: 'Failed to resume cortex',
    });
  }
});

/**
 * GET /api/v1/cortex/feedback
 * Get active/used providers and their feedback metrics
 * Returns only providers starting with 'non' prefix
 */
router.get('/feedback', (_req: Request, res: Response) => {
  try {
    const providers = cortex.providerFeedbackEngine.getSerializableFeedback();
    const activeProvider = cortex.providerFeedbackEngine.getActiveProvider();

    res.json({
      ok: true,
      providers,
      activeProvider: activeProvider
        ? {
            name: activeProvider.name,
            provider: activeProvider.provider,
            status: activeProvider.status,
            isActive: activeProvider.status === 'processing',
          }
        : null,
    });
  } catch (e) {
    console.error('[Cortex Routes] Feedback fetch error:', e);
    res.status(500).json({
      ok: false,
      error: 'Failed to fetch provider feedback',
    });
  }
});

/**
 * GET /api/v1/cortex/session/:sessionId
 * Get session context and highlights from current session
 */
router.get('/session/:sessionId', (req: Request, res: Response) => {
  try {
    const { sessionId } = req.params;
    if (!sessionId) {
      return res.status(400).json({ ok: false, error: 'Missing sessionId' });
    }
    const context = cortex.sessionContextService.getSerializableContext(sessionId);

    res.json({
      ok: true,
      context,
    });
  } catch (e) {
    console.error('[Cortex Routes] Session context fetch error:', e);
    res.status(500).json({
      ok: false,
      error: 'Failed to fetch session context',
    });
  }
});

/**
 * GET /api/v1/cortex/memory/:sessionId
 * Get auto-filled short and long-term memory for session
 */
router.get('/memory/:sessionId', async (req: Request, res: Response) => {
  try {
    const { sessionId } = req.params;
    if (!sessionId) {
      return res.status(400).json({ ok: false, error: 'Missing sessionId' });
    }

    // Force memory update if requested
    const forceUpdate = req.query['force'] === 'true';
    if (forceUpdate) {
      await cortex.memoryAutoFiller.forceUpdate(sessionId);
    }

    const memory = cortex.memoryAutoFiller.getSerializableMemory(sessionId);

    res.json({
      ok: true,
      memory,
    });
  } catch (e) {
    console.error('[Cortex Routes] Memory fetch error:', e);
    res.status(500).json({
      ok: false,
      error: 'Failed to fetch auto-filled memory',
    });
  }
});

/**
 * POST /api/v1/cortex/session/:sessionId/highlight
 * Manually add a highlight to session
 */
router.post('/session/:sessionId/highlight', (req: Request, res: Response) => {
  try {
    const { sessionId } = req.params;
    if (!sessionId) {
      return res.status(400).json({ ok: false, error: 'Missing sessionId' });
    }
    const { type, content, icon } = req.body;

    if (!type || !content) {
      res.status(400).json({
        ok: false,
        error: 'Missing required fields: type, content',
      });
      return;
    }

    cortex.sessionContextService.addHighlight(sessionId, {
      type,
      content,
      relevanceScore: 0.8,
      icon,
    });

    res.json({
      ok: true,
      message: 'Highlight added',
    });
  } catch (e) {
    console.error('[Cortex Routes] Highlight add error:', e);
    res.status(500).json({
      ok: false,
      error: 'Failed to add highlight',
    });
  }
});

/**
 * POST /api/v1/cortex/session/:sessionId/goal
 * Set current session goal
 */
router.post('/session/:sessionId/goal', (req: Request, res: Response) => {
  try {
    const { sessionId } = req.params;
    if (!sessionId) {
      return res.status(400).json({ ok: false, error: 'Missing sessionId' });
    }
    const { goal } = req.body;

    if (!goal) {
      res.status(400).json({
        ok: false,
        error: 'Missing required field: goal',
      });
      return;
    }

    cortex.sessionContextService.setCurrentGoal(sessionId, goal);

    res.json({
      ok: true,
      message: 'Goal set',
    });
  } catch (e) {
    console.error('[Cortex Routes] Goal set error:', e);
    res.status(500).json({
      ok: false,
      error: 'Failed to set goal',
    });
  }
});

/**
 * POST /api/v1/cortex/memory/:sessionId/update
 * Manually update auto-filled memory
 */
router.post('/memory/:sessionId/update', async (req: Request, res: Response) => {
  try {
    const { sessionId } = req.params;
    if (!sessionId) {
      return res.status(400).json({ ok: false, error: 'Missing sessionId' });
    }
    await cortex.memoryAutoFiller.forceUpdate(sessionId);

    const memory = cortex.memoryAutoFiller.getSerializableMemory(sessionId);

    res.json({
      ok: true,
      memory,
    });
  } catch (e) {
    console.error('[Cortex Routes] Memory update error:', e);
    res.status(500).json({
      ok: false,
      error: 'Failed to update memory',
    });
  }
});

/**
 * GET /api/v1/cortex/providers
 * Get all active providers (alternative endpoint)
 */
router.get('/providers', (_req: Request, res: Response) => {
  try {
    const providers = cortex.providerFeedbackEngine.getActiveProviders();

    res.json({
      ok: true,
      providers: providers.map((p) => ({
        name: p.name,
        provider: p.provider,
        status: p.status,
        callCount: p.callCount,
        successRate: p.successRate,
        avgLatencyMs: p.avgLatencyMs,
        confidenceScore: p.confidenceScore,
        lastUsed: p.lastUsed.toISOString(),
      })),
    });
  } catch (e) {
    console.error('[Cortex Routes] Providers fetch error:', e);
    res.status(500).json({
      ok: false,
      error: 'Failed to fetch providers',
    });
  }
});

/**
 * POST /api/v1/cortex/providers/reset
 * Reset provider feedback for new session
 */
router.post('/providers/reset', (_req: Request, res: Response) => {
  try {
    cortex.providerFeedbackEngine.resetSession();

    res.json({
      ok: true,
      message: 'Provider session reset',
    });
  } catch (e) {
    console.error('[Cortex Routes] Provider reset error:', e);
    res.status(500).json({
      ok: false,
      error: 'Failed to reset providers',
    });
  }
});

export const cortexRoutes = router;